# [patch] Speed up JIT time by removing unnecessary for / list comprehension