Tries to address #12 partially.

At cf9dfee617bd337545a9e9d70a7709b9269ca5d4 (i.e., before #1 ), it would take around ~9s to JIT compile for collision-aware IK with yumi and panda; at b192178d62e86d12ce1c0d1d4a22cc853156c056 (~main as of 12/8) it takes around 61s to JIT panda, and 18 seconds for yumi.

With this push, compilation time is brought down to ~20s for panda and ~16s for yumi.

In general, there's a lot of code inefficiencies to be addressed, esp with control workflows.